# docs/ui/control-plane-mock/openapi.yaml
openapi: 3.1.0
info:
  title: Quant Research Platform Control Plane UI Mock
  version: 0.2.0
  description: >
    G0 frontend contract baseline. Endpoints marked frontend-proposal require
    backend review before implementation.
  license:
    name: Proprietary
servers:
  - url: http://localhost:4010
    description: Local mock
security:
  - oidc: []
  - bearerAuth: []
...
paths:
...
  /v1/research-jobs:
    get:
      tags: [ResearchJobs]
      operationId: listResearchJobs
      summary: List authorized research jobs
      x-source: frontend-proposal
      parameters:
        - $ref: "#/components/parameters/Market"
        - $ref: "#/components/parameters/State"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Paginated research jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchJobPage"
    post:
      tags: [ResearchJobs]
      operationId: createResearchJob
      summary: Create a research job
      x-source: technical-design
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResearchJobCommand"
      responses:
        "202":
          $ref: "#/components/responses/CommandAccepted"
        "400":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/research-jobs/{job_id}:
    get:
      tags: [ResearchJobs]
      operationId: getResearchJob
      summary: Get a research job snapshot
      x-source: technical-design
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Current server snapshot
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchJob"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/research-jobs/{job_id}/brief-versions:
    get:
      tags: [ResearchJobs]
      operationId: listResearchBriefVersions
      summary: List research brief versions
      x-source: g0-contract-baseline
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Immutable brief version history, including drafts
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ResearchBriefVersion"
                  next_cursor:
                    type: [string, "null"]
        "404":
          $ref: "#/components/responses/Problem"
    post:
      tags: [ResearchJobs]
      operationId: createResearchBriefVersion
      summary: Create a draft research brief version
      x-source: g0-contract-baseline
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResearchBriefVersionCommand"
      responses:
        "202":
          $ref: "#/components/responses/CommandAccepted"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/research-brief-versions/{brief_version_id}:
    get:
      tags: [ResearchJobs]
      operationId: getResearchBriefVersion
      summary: Get a research brief version
      x-source: g0-contract-baseline
      parameters:
        - $ref: "#/components/parameters/BriefVersionId"
      responses:
        "200":
          description: One research brief version
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchBriefVersion"
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      tags: [ResearchJobs]
      operationId: updateResearchBriefVersion
      summary: Update a draft research brief version
      x-source: g0-contract-baseline
      parameters:
        - $ref: "#/components/parameters/BriefVersionId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResearchBriefVersionCommand"
      responses:
        "202":
          $ref: "#/components/responses/CommandAccepted"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/research-brief-versions/{brief_version_id}:freeze:
    post:
      tags: [ResearchJobs]
      operationId: freezeResearchBriefVersion
      summary: Freeze a research brief version
      x-source: g0-contract-baseline
      parameters:
        - $ref: "#/components/parameters/BriefVersionId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandMetadata"
      responses:
        "202":
          $ref: "#/components/responses/CommandAccepted"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag of the aggregate version being changed.
      schema:
        type: string
        minLength: 3
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
...
    BriefVersionId:
      name: brief_version_id
      in: path
      required: true
      schema:
        type: string
    Market:
      name: market
      in: query
      schema:
        $ref: "#/components/schemas/MarketId"
    State:
      name: state
      in: query
      schema:
        $ref: "#/components/schemas/ResearchJobState"
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    CommandAccepted:
      description: Command accepted for asynchronous processing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CommandReceipt"
    Problem:
      description: RFC 9457-style problem response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    MarketId:
      type: string
      enum: [CN_A, CN_COMMODITY_FUTURES]
    Environment:
      type: string
      enum: [RESEARCH, PAPER, LIVE]
    FrequencyId:
      type: string
      enum: [1d]
    ResearchJobState:
      type: string
      enum:
        - DRAFT
        - READY
        - RUNNING
        - WAITING_INPUT
        - BLOCKED_POLICY
        - SUCCEEDED
        - FAILED
        - CANCELLED
        - ARCHIVED
    ResearchBriefVersionState:
      type: string
      enum: [DRAFT, FROZEN, SUPERSEDED]
...
    AttemptState:
      type: string
      enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED, TIMED_OUT]
...
    Budget:
      type: object
      additionalProperties: false
      required: [candidate_limit, wall_clock_minutes]
      properties:
        candidate_limit:
          type: integer
          minimum: 1
        llm_token_limit:
          type: integer
          minimum: 0
        cpu_hours:
          type: number
          minimum: 0
        wall_clock_minutes:
          type: integer
          minimum: 1
    CommandMetadata:
      type: object
      additionalProperties: false
      required: [reason, parent_artifact_id, budget, schema_version]
      properties:
        reason:
          type: string
          minLength: 3
        parent_artifact_id:
          type: [string, "null"]
        budget:
          $ref: "#/components/schemas/Budget"
        schema_version:
          type: string
    CreateResearchJobCommand:
      type: object
      additionalProperties: false
      required:
        - metadata
        - market
        - universe_ref
        - frequency
        - decision_clock
        - trade_clock
        - horizon
        - research_brief_version_id
      properties:
        metadata:
          $ref: "#/components/schemas/CommandMetadata"
        market:
          $ref: "#/components/schemas/MarketId"
        universe_ref:
          type: string
        frequency:
          $ref: "#/components/schemas/FrequencyId"
        decision_clock:
          type: string
        trade_clock:
          type: string
        settlement_clock:
          type: string
        exchange_scope:
          type: array
          minItems: 1
          items:
            type: string
            enum: [SHFE, INE, DCE, CZCE, GFEX]
        contract_selection:
          type: string
          enum: [ACTUAL_CONTRACTS_ONLY]
        roll_policy:
          type: string
          description: Immutable RollPolicyVersion reference.
        horizon:
          type: string
        research_brief_version_id:
          type: string
      allOf:
        - if:
            properties:
              market:
                const: CN_COMMODITY_FUTURES
            required: [market]
          then:
            properties:
              settlement_clock:
                type: string
              exchange_scope:
                type: array
                minItems: 1
                items:
                  type: string
              contract_selection:
                type: string
                const: ACTUAL_CONTRACTS_ONLY
              roll_policy:
                type: string
            required:
              - settlement_clock
              - exchange_scope
              - contract_selection
              - roll_policy
    ResearchBriefContent:
      type: object
      required:
        - hypothesis
        - economic_mechanism
        - expected_direction
        - falsification_conditions
        - allowed_data_domains
        - constraints
      properties:
        hypothesis:
          type: string
          minLength: 3
        economic_mechanism:
          type: string
          minLength: 3
        expected_direction:
          type: string
          enum: [POSITIVE, NEGATIVE, NON_MONOTONIC, UNKNOWN]
        falsification_conditions:
          type: array
          minItems: 1
          items:
            type: string
        allowed_data_domains:
          type: array
          items:
            type: string
        forbidden_data_domains:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        evidence_ref_ids:
          type: array
          items:
            type: string
        uncertainties:
          type: array
          items:
            type: string
    ResearchBriefVersion:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/ResearchBriefContent"
        - type: object
          required:
            - id
            - job_id
            - version
            - resource_version
            - status
            - content_hash
            - created_at
            - created_by
          properties:
            id:
              type: string
            job_id:
              type: string
            version:
              type: integer
              minimum: 1
            resource_version:
              type: integer
              minimum: 1
            status:
              $ref: "#/components/schemas/ResearchBriefVersionState"
            content_hash:
              type: [string, "null"]
            created_at:
              type: string
              format: date-time
            created_by:
              type: string
            frozen_at:
              type: [string, "null"]
              format: date-time
            frozen_by:
              type: [string, "null"]
    CreateResearchBriefVersionCommand:
      type: object
      additionalProperties: false
      required: [metadata, brief]
      properties:
        metadata:
          $ref: "#/components/schemas/CommandMetadata"
        brief:
          $ref: "#/components/schemas/ResearchBriefContent"
    UpdateResearchBriefVersionCommand:
      type: object
      additionalProperties: false
      required: [metadata, brief]
      properties:
        metadata:
          $ref: "#/components/schemas/CommandMetadata"
        brief:
          $ref: "#/components/schemas/ResearchBriefContent"
    ResourceFreshness:
      type: object
      required: [as_of, is_stale]
      properties:
        as_of:
          type: string
          format: date-time
        is_stale:
          type: boolean
        stale_reason:
          type: [string, "null"]
    ResearchJob:
      type: object
      required:
        - id
        - version
        - title
        - market
        - environment
        - state
        - owner
        - budget
        - freshness
        - allowed_actions
      properties:
        id:
          type: string
        version:
          type: string
        title:
          type: string
        market:
          $ref: "#/components/schemas/MarketId"
        environment:
          $ref: "#/components/schemas/Environment"
        state:
          $ref: "#/components/schemas/ResearchJobState"
        owner:
          type: string
        current_stage:
          type: string
        budget:
          $ref: "#/components/schemas/Budget"
        budget_used:
          type: object
          additionalProperties:
            type: number
        latest_attempt:
          $ref: "#/components/schemas/Attempt"
        snapshot_refs:
          type: array
          items:
            type: string
        policy_version:
          type: string
        run_fingerprint:
          type: [string, "null"]
        freshness:
          $ref: "#/components/schemas/ResourceFreshness"
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/Blocker"
        allowed_actions:
          type: array
          items:
            type: string
        updated_at:
          type: string
          format: date-time
    ResearchJobPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResearchJob"
        next_cursor:
          type: [string, "null"]
    Attempt:
      type: object
      required: [attempt, state]
      properties:
        id:
          type: string
        attempt:
          type: integer
          minimum: 1
        state:
          $ref: "#/components/schemas/AttemptState"
        started_at:
          type: [string, "null"]
          format: date-time
        heartbeat_at:
          type: [string, "null"]
          format: date-time
        error_code:
          type: [string, "null"]
    Blocker:
      type: object
      required: [code, title, responsibility]
      properties:
        code:
          type: string
        title:
          type: string
        detail:
          type: string
        responsibility:
          enum: [RESEARCHER, RESEARCH_LEAD, RISK_REVIEWER, DATA_STEWARD, PLATFORM_ADMIN]
        policy_version:
          type: [string, "null"]
...
    CommandReceipt:
      type: object
      required: [command_id, status, resource_id, submitted_at]
      properties:
        command_id:
          type: string
        status:
          const: ACCEPTED
        resource_id:
          type: string
        submitted_at:
          type: string
          format: date-time
    Problem:
      type: object
      required: [type, title, status, code, request_id, retryable]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        current_version:
          type: [string, "null"]
        field_errors:
          type: array
          items:
            type: object
            required: [path, code, message]
            properties:
              path:
                type: string
              code:
                type: string
              message:
                type: string